+++ src/Symfony/Component/Security/Http/CHANGELOG.md
@@ -7,6 +7,10 @@ CHANGELOG
  * Add `RememberMeBadge` to `JsonLoginAuthenticator` and enable reading parameter in JSON request body
  * Add argument `$exceptionCode` to `#[IsGranted]`
  * Deprecate passing a secret as the 2nd argument to the constructor of `Symfony\Component\Security\Http\RememberMe\PersistentRememberMeHandler`
+ * Add `OidcUserInfoTokenHandler` and `OidcTokenHandler` with OIDC support for `AccessTokenAuthenticator`
+ * Add `attributes` optional array argument in `UserBadge`
+ * Call `UserBadge::userLoader` with attributes if the argument is set
+ * Allow to override badge fqcn on `Passport::addBadge`
 
 6.2
 ---